feat(polka-storage-provider-client): simplify the client command
