+++ cli/polka-storage-provider/client/src/commands/mod.rs
@@ -1,11 +1,22 @@
-mod client;
 mod proofs;
 mod wallet;
 
 use clap::Parser;
+use jsonrpsee::core::ClientError;
+use polka_storage_provider_common::rpc::StorageProviderRpcClient;
+use storagext::{
+    deser::DeserializablePath,
+    multipair::{MultiPairArgs, MultiPairSigner},
+    types::market::{ClientDealProposal as SxtClientDealProposal, DealProposal as SxtDealProposal},
+};
+use url::Url;
 
-use self::{client::ClientCommand, proofs::ProofsCommand};
+use self::proofs::ProofsCommand;
 pub(super) use crate::commands::wallet::WalletCommand;
+use crate::rpc_client::PolkaStorageRpcClient;
+
+/// Default RPC server's URL.
+const DEFAULT_RPC_SERVER_URL: &str = "http://127.0.0.1:8000";
 
 /// CLI components error handling implementor.
 #[derive(Debug, thiserror::Error)]
@@ -29,10 +40,13 @@ pub enum CliError {
     MaterError(#[from] mater::Error),
 
     #[error(transparent)]
-    RpcCommand(#[from] crate::commands::client::ClientCommandError),
-
-    #[error(transparent)]
     UtilsCommand(#[from] crate::commands::proofs::UtilsCommandError),
+
+    #[error("the RPC client failed: {0}")]
+    RpcClient(#[from] ClientError),
+
+    #[error("no signer key was provider")]
+    NoSigner,
 }
 
 /// A CLI application that facilitates management operations over a running full
@@ -44,12 +58,46 @@ pub(crate) enum Cli {
     #[command(subcommand)]
     Wallet(WalletCommand),
 
-    /// Client RPC commands.
-    Client(ClientCommand),
-
     /// Utility commands for storage related actions.
     #[command(subcommand)]
     Proofs(ProofsCommand),
+
+    /// Retrieve information about the provider's node.
+    Info {
+        /// URL of the providers RPC server.
+        #[arg(long, default_value = DEFAULT_RPC_SERVER_URL)]
+        rpc_server_url: Url,
+    },
+
+    /// Propose a storage deal.
+    ProposeDeal {
+        /// URL of the providers RPC server.
+        #[arg(long, default_value = DEFAULT_RPC_SERVER_URL)]
+        rpc_server_url: Url,
+        /// Storage deal to propose. Either JSON or a file path, prepended with an @.
+        #[arg(value_parser = <SxtDealProposal as DeserializablePath>::deserialize_json )]
+        deal_proposal: SxtDealProposal,
+    },
+
+    /// Publish a signed storage deal.
+    PublishDeal {
+        /// URL of the providers RPC server.
+        #[arg(long, default_value = DEFAULT_RPC_SERVER_URL)]
+        rpc_server_url: Url,
+        /// Storage deal to publish. Either JSON or a file path, prepended with an @.
+        #[arg(value_parser = <SxtClientDealProposal as DeserializablePath>::deserialize_json)]
+        client_deal_proposal: SxtClientDealProposal,
+    },
+
+    /// Sign a storage deal using the provided key, will output the deal as a JSON
+    /// — no information is shared across the network.
+    SignDeal {
+        #[arg(value_parser = <SxtDealProposal as DeserializablePath>::deserialize_json)]
+        deal_proposal: SxtDealProposal,
+
+        #[command(flatten)]
+        signer_key: MultiPairArgs,
+    },
 }
 
 impl Cli {
@@ -69,8 +117,69 @@ impl Cli {
                 WalletCommand::Verify(cmd) => Ok(cmd.run()?),
                 WalletCommand::Sign(cmd) => Ok(cmd.run()?),
             },
-            Self::Client(cmd) => Ok(cmd.run().await?),
             Self::Proofs(utils) => Ok(utils.run().await?),
+            Self::Info { rpc_server_url } => Self::info(rpc_server_url).await,
+            Self::ProposeDeal {
+                rpc_server_url,
+                deal_proposal,
+            } => Self::propose_deal(rpc_server_url, deal_proposal).await,
+            Self::PublishDeal {
+                rpc_server_url,
+                client_deal_proposal,
+            } => Self::publish_deal(rpc_server_url, client_deal_proposal).await,
+            Self::SignDeal {
+                deal_proposal,
+                signer_key,
+            } => Self::sign_deal(deal_proposal, signer_key),
         }
+    }
+
+    async fn info(rpc_server_url: Url) -> Result<(), CliError> {
+        let client = PolkaStorageRpcClient::new(&rpc_server_url).await?;
+        let info = client.info().await?;
+        println!(
+            "{}",
+            serde_json::to_string_pretty(&info)
+                .expect("type is serializable so this call should never fail")
+        );
+        Ok(())
+    }
+
+    async fn propose_deal(
+        rpc_server_url: Url,
+        deal_proposal: SxtDealProposal,
+    ) -> Result<(), CliError> {
+        let client = PolkaStorageRpcClient::new(&rpc_server_url).await?;
+        let result = client.propose_deal(deal_proposal).await?;
+        println!("{}", result);
+        Ok(())
+    }
+
+    async fn publish_deal(
+        rpc_server_url: Url,
+        client_deal_proposal: SxtClientDealProposal,
+    ) -> Result<(), CliError> {
+        let client = PolkaStorageRpcClient::new(&rpc_server_url).await?;
+        let result = client.publish_deal(client_deal_proposal).await?;
+        println!("{}", result);
+        Ok(())
+    }
+
+    fn sign_deal(
+        deal_proposal: SxtDealProposal,
+        signer_key: MultiPairArgs,
+    ) -> Result<(), CliError> {
+        let Some(signer) = Option::<MultiPairSigner>::from(signer_key) else {
+            return Err(CliError::NoSigner);
+        };
+
+        let signature = deal_proposal.sign_serializable(&signer);
+
+        println!(
+            "{}",
+            serde_json::to_string_pretty(&signature)
+                .expect("the type is serializable, so this should never fail")
+        );
+        Ok(())
     }
 }
