--- docs/src/storage-provider-cli/client/client.md
@@ -1,10 +1,0 @@
-# `client`
-
-<!-- TODO: add more info / parameters -->
-
-| Name         | Description                                                                                                              |
-| ------------ | ------------------------------------------------------------------------------------------------------------------------ |
-| info         | Retrieve information about the provider's node                                                                           |
-| propose-deal | Propose a storage deal                                                                                                   |
-| publish-deal | Publish a signed storage deal                                                                                            |
-| sign-deal    | Sign a storage deal using the provided key, will output the deal as a JSON — no information is shared across the network |
